requirements.txt
pytube
openai-whisper
torch
ffmpeg-python
